開発用データベース変更

+++ db/migrate/20241021164937_change_end_duration_type_in_boards.rb
@@ -1,25 +1,33 @@
 class ChangeEndDurationTypeInBoards < ActiveRecord::Migration[7.0]
   def up
-    add_column :boards, :end_duration_temp, :timestamp_
-    Board.reset_column_information
-    Board.find_each do |board|
-      if board.end_duration.present?
-        board.update_column(:end_duration_temp, board.end_duration.to_datetime)
+    if ActiveRecord::Base.connection.adapter_name == 'PostgreSQL'
+      add_column :boards, :end_duration_temp, :timestamp
+      
+      Board.reset_column_information
+      Board.find_each do |board|
+        if board.end_duration.present?
+          board.update_column(:end_duration_temp, board.end_duration.to_datetime)
+        end
       end
+
+      remove_column :boards, :end_duration
+      rename_column :boards, :end_duration_temp, :end_duration
     end
-    remove_column :boards, :end_duration
-    rename_column :boards, :end_duration_temp, :end_duration
   end
 
   def down
-    add_column :boards, :end_duration_temp, :date
-    Board.reset_column_information
-    Board.find_each do |board|
-      if board.end_duration.present?
-        board.update_column(:end_duration_temp, board.end_duration.to_date)
+    if ActiveRecord::Base.connection.adapter_name == 'PostgreSQL'
+      add_column :boards, :end_duration_temp, :integer
+      
+      Board.reset_column_information
+      Board.find_each do |board|
+        if board.end_duration.present?
+          board.update_column(:end_duration_temp, board.end_duration.to_i)
+        end
       end
+
+      remove_column :boards, :end_duration
+      rename_column :boards, :end_duration_temp, :end_duration
     end
-    remove_column :boards, :end_duration
-    rename_column :boards, :end_duration_temp, :end_duration
   end
 end
